Completed refactoring Navbar and components
