--- frontend/src/app/userDashboard/navbarContainer/components/Navbar.js
@@ -1,104 +1,0 @@
-import React, { useState, useRef } from "react";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import IconButton from "@material-ui/core/IconButton";
-import Typography from "@material-ui/core/Typography";
-import InputBase from "@material-ui/core/InputBase";
-import Badge from "@material-ui/core/Badge";
-import MenuIcon from "@material-ui/icons/Menu";
-import SearchIcon from "@material-ui/icons/Search";
-import AccountCircle from "@material-ui/icons/AccountCircle";
-import MailIcon from "@material-ui/icons/Mail";
-import NotificationsIcon from "@material-ui/icons/Notifications";
-import MoreIcon from "@material-ui/icons/MoreVert";
-import { Logo } from "./logo";
-
-export const Navbar = ({
-  classes,
-  user,
-  profileMenuId,
-  handleProfileMenuOpen,
-  mobileMenuId,
-  handleMobileMenuOpen,
-}) => {
-  const [search, setSearch] = useState("");
-  const searchAnchor = useRef(null);
-
-  return (
-    <AppBar position="static">
-      <Toolbar>
-        <IconButton
-          edge="start"
-          className={classes.menuButton}
-          color="inherit"
-          aria-label="open drawer"
-        >
-          <MenuIcon />
-        </IconButton>
-        <Typography className={classes.title} variant="h6" noWrap>
-          <Logo />
-        </Typography>
-        <div className={classes.search}>
-          <div className={classes.searchIcon}>
-            <SearchIcon />
-          </div>
-          <InputBase
-            placeholder="Google Search…"
-            classes={{
-              root: classes.inputRoot,
-              input: classes.inputInput,
-            }}
-            inputProps={{ "aria-label": "search" }}
-            value={search}
-            onChange={(e) => setSearch(e.target.value)}
-            onKeyUp={(e) => {
-              if (e.key === "Enter") {
-                searchAnchor.current.click();
-              }
-            }}
-          />
-          <a
-            href={"https://www.google.com/search?&q=" + search}
-            target="_blank"
-            rel="noopener"
-            ref={searchAnchor}
-          />
-        </div>
-        <div className={classes.grow} />
-        <div className={classes.sectionDesktop}>
-          <IconButton aria-label="show 4 new messagess" color="inherit">
-            <Badge badgeContent={user.messages?.length} color="secondary">
-              <MailIcon />
-            </Badge>
-          </IconButton>
-          <IconButton aria-label="show 17 new notifications" color="inherit">
-            <Badge badgeContent={user.notifications?.length} color="secondary">
-              <NotificationsIcon />
-            </Badge>
-          </IconButton>
-          <IconButton
-            edge="end"
-            aria-label="account of current user"
-            aria-controls={profileMenuId}
-            aria-haspopup="true"
-            onClick={handleProfileMenuOpen}
-            color="inherit"
-          >
-            <AccountCircle />
-          </IconButton>
-        </div>
-        <div className={classes.sectionMobile}>
-          <IconButton
-            aria-label="show more"
-            aria-controls={mobileMenuId}
-            aria-haspopup="true"
-            onClick={handleMobileMenuOpen}
-            color="inherit"
-          >
-            <MoreIcon />
-          </IconButton>
-        </div>
-      </Toolbar>
-    </AppBar>
-  );
-};
