+++ frontend/src/app/userDashboard/navbarContainer/navbar/sections/index.js
@@ -1,0 +1,3 @@
+export * from "./Desktop";
+export * from "./Mobile";
+export * from "./Search";
